Split interaction macros into separate files
The macros file within interactions has grown considerably so this
change splits them into smaller, more maintainable files.
