+++ src/apps/interactions/macros/collection-sort-form.js
@@ -1,0 +1,26 @@
+module.exports = {
+  method: 'get',
+  class: 'c-collection__sort-form js-AutoSubmit',
+  hideFormActions: true,
+  hiddenFields: { custom: true },
+  children: [
+    {
+      macroName: 'MultipleChoiceField',
+      label: 'Sort by',
+      name: 'sortby',
+      modifier: ['small', 'inline', 'light'],
+      options: [
+        { value: 'date:desc', label: 'Newest' },
+        { value: 'date:asc', label: 'Oldest' },
+        { value: 'company.name:asc', label: 'Company: A-Z' },
+        { value: 'company.name:desc', label: 'Company: Z-A' },
+        { value: 'contact.name:asc', label: 'Contact: A-Z' },
+        { value: 'contact.name:desc', label: 'Contact: Z-A' },
+        { value: 'dit_adviser.name:asc', label: 'Adviser: A-Z' },
+        { value: 'dit_adviser.name:desc', label: 'Adviser: Z-A' },
+        { value: 'subject:asc', label: 'Subject: A-Z' },
+        { value: 'subject:desc', label: 'Subject: Z-A' },
+      ],
+    },
+  ],
+}
