+++ src/apps/interactions/macros/index.js
@@ -1,0 +1,15 @@
+const collectionSortForm = require('./collection-sort-form')
+const collectionFilterFields = require('./collection-filter-fields')
+const fields = require('./fields')
+const interactionForm = require('./interaction-form')
+const kindForm = require('./kind-form')
+const serviceDeliveryForm = require('./service-delivery-form')
+
+module.exports = {
+  collectionSortForm,
+  collectionFilterFields,
+  fields,
+  interactionForm,
+  kindForm,
+  serviceDeliveryForm,
+}
